refactor: 프로필 사진 변경 컴포넌트 분리
